corsika limits [skip ci]
